refactor(ticket): split ticket persistence repositories
Summary:
- split ticket persistence into activity, query, command, link,
  comment, runtime, and usage repositories under
  `internal/repo/ticket`
- change `internal/ticket` to depend on smaller repository
  interfaces via explicit dependency wiring instead of a single
  all-purpose repo
- add direct repo integration tests for CRUD, comment revisions,
  dependencies and external links, pickup diagnosis, and runtime
  query behavior

Rationale:
- narrow repository boundaries so CRUD, comments, links, and
  runtime diagnosis can evolve without colliding in one large file
- make service wiring show ownership of each ticket subdomain
  directly instead of hiding responsibility behind a broad facade
- improve regression isolation with repo-level tests that exercise
  the new boundaries without going through higher layers

Tests:
- go test ./internal/repo/ticket -count=1
- go test ./internal/repo/... ./internal/ticket ./internal/httpapi
