+++ internal/orchestrator/scheduler.go
@@ -68,7 +68,15 @@ func NewScheduler(client *ent.Client, logger *slog.Logger, events provider.Event
 		client:        client,
 		logger:        logger.With("component", "scheduler"),
 		events:        events,
-		scheduledJobs: scheduledjobservice.NewService(scheduledjobrepo.NewEntRepository(client), ticketservice.NewService(ticketrepo.NewEntRepository(client)), logger),
+		scheduledJobs: scheduledjobservice.NewService(scheduledjobrepo.NewEntRepository(client), ticketservice.NewService(ticketservice.Dependencies{
+			Activity: ticketrepo.NewActivityRepository(client),
+			Query:    ticketrepo.NewQueryRepository(client),
+			Command:  ticketrepo.NewCommandRepository(client),
+			Link:     ticketrepo.NewLinkRepository(client),
+			Comment:  ticketrepo.NewCommentRepository(client),
+			Usage:    ticketrepo.NewUsageRepository(client),
+			Runtime:  ticketrepo.NewRuntimeRepository(client),
+		}), logger),
 		now:           time.Now,
 	}
 }
