+++ internal/repo/ticket/pg_fixture_test.go
@@ -1,0 +1,16 @@
+package ticket
+
+import (
+	"os"
+	"testing"
+
+	"github.com/BetterAndBetterII/openase/internal/testutil/pgtest"
+)
+
+var testPostgres *pgtest.Server
+
+func TestMain(m *testing.M) {
+	os.Exit(pgtest.RunTestMain(m, "repo_ticket", func(server *pgtest.Server) {
+		testPostgres = server
+	}))
+}
