+++ internal/ticket/test_helpers_test.go
@@ -12,5 +12,13 @@ func newTicketStatusService(client *ent.Client) *ticketstatus.Service {
 }
 
 func newTicketService(client *ent.Client) *Service {
-	return NewService(ticketrepo.NewEntRepository(client))
+	return NewService(Dependencies{
+		Activity: ticketrepo.NewActivityRepository(client),
+		Query:    ticketrepo.NewQueryRepository(client),
+		Command:  ticketrepo.NewCommandRepository(client),
+		Link:     ticketrepo.NewLinkRepository(client),
+		Comment:  ticketrepo.NewCommentRepository(client),
+		Usage:    ticketrepo.NewUsageRepository(client),
+		Runtime:  ticketrepo.NewRuntimeRepository(client),
+	})
 }
